add identity auth

+++ controller/StockController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.Dtos.Stock;
+using api.Helpers;
 using api.interfaces;
 using api.Mappers;
 using Microsoft.AspNetCore.Mvc;
@@ -13,7 +14,7 @@ namespace api.controller
 {
     [Route("api/stock")]
     [ApiController]
-    public class StockController: ControllerBase
+    public class StockController : ControllerBase
     {
         private readonly ApplicationDBContext _context;
         private readonly IstockRepository _stockRepo;
@@ -23,15 +24,19 @@ namespace api.controller
             _stockRepo = stockRepo;
         }
         [HttpGet]
-        public async Task<IActionResult>  GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
-            var stock = await _stockRepo.GetAllAsync();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var stock = await _stockRepo.GetAllAsync(query);
             var stockDto = stock.Select(x => x.ToStockDto());
             return Ok(stock);
         }
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if (stock == null)
@@ -41,29 +46,35 @@ namespace api.controller
             return Ok(stock.ToStockDto());
         }
         [HttpPost]
-        public async Task<IActionResult>  Create([FromBody] CreateStockRequestDtos stockDto)
+        public async Task<IActionResult> Create([FromBody] CreateStockRequestDtos stockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = stockDto.ToStockFromCreateDto();
             await _stockRepo.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
-        public  async Task<IActionResult>  Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
+        [Route("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
-            var stockModel = await  _stockRepo.UpdateAsync(id, updateDto);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             if (stockModel == null)
             {
                 return NotFound();
             }
-            
+
             return Ok(stockModel.ToStockDto());
         }
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = await _stockRepo.DeleteAsync(id);
             if (stockModel == null)
             {
@@ -72,6 +83,6 @@ namespace api.controller
             return Ok();
         }
     }
-  
-     
+
+
 }
